requirements.txt
jbtech-utils-package
email-controller
flask
psycopg2